And now a warning generated by g++ that clang++ missed.
git-svn-id: svn://svn.code.sf.net/p/stella/code/trunk@2721 8b62c5a3-ac7e-4cc8-8f21-d9a121418aba

+++ src/gui/Dialog.cxx
@@ -705,7 +705,7 @@ void Dialog::TabFocus::appendFocusList(WidgetArray& list)
 {
   int active = widget->getActiveTab();
 
-  if(active >= 0 && active < focus.size())
+  if(active >= 0 && active < (int)focus.size())
     list.push_back(focus[active].list);
 }
